- Warning fix :)
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@29450 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/tools/lighter2/lighter.cpp
@@ -122,7 +122,7 @@ namespace lighter
           maxSwapSize = 200;
       }
       // Check for override.
-      maxSwapSize = (size_t)(configMgr->GetInt ("lighter2.swapcachesize", maxSwapSize)*1024*1024);
+      maxSwapSize = (size_t)(configMgr->GetInt ("lighter2.swapcachesize", (int)maxSwapSize)*1024*1024);
       swapManager = new SwapManager (maxSwapSize);
     }
